Change import- and start  logic, removed cli.py
