+++ tests/test_bitcoind.py
@@ -3,7 +3,7 @@ import logging
 
 def test_bitcoinddocker_running(caplog):
     caplog.set_level(logging.DEBUG)
-    from bitcoind import BitcoindDockerController
+    from specter.bitcoind import BitcoindDockerController
     my_bitcoind = BitcoindDockerController(rpcport=18999) # completly different port to not interfere
     #assert my_bitcoind.detect_bitcoind_container() == True
     rpcconn = my_bitcoind.start_bitcoind(cleanup_at_exit=True)
