+++ tests/test_cli.py
@@ -1,4 +1,4 @@
-from cli import fetch_wallet_addresses_for_mining
+from specter.bitcoind import fetch_wallet_addresses_for_mining
 
 def test_fetch_wallet_addresses_for_mining(wallets_filled_data_folder):
     # Todo: instantiate a specter-testwallet
